created elevator branch and transferred code from robot2025
updated elevator--setting position not speed

restructured code for maintaining elevator height; removed inches because inches to ticks conversion was wrong

testing by changing pid values and output denominator; changed logic by taking in desiredChange rather than desiredTicks

theoretical feed forward

added integrated feedforward and profiledPID

added tunable values

elevator testing - moving but not stopping at setpoint - see updates 5/28

lots of code cleaning - fixed most of the logic: elevator command now functions like an instant command; elevator is moving and stopping now with error

put the pid controller calculate part before the feedforward calculate method because apparently the pid calculator also incrememnts the trapezoid profile so it has to be before... also changed the pid calculator to take the current position as an input rather than the goal position because this is what the definition says it needs

TUNABLES WORK (added in Logger.start() in Robot() also there appear to be changes in BuildConstants and I don't know why, also the computer started beeping like it was going to explode (ticking bomb vibes) while I was trying to push, not sure it that's relevant?

got feedforward mostly tuned, added a elevator stop method that we call in disabled periodic so now it doesn't jump when it reenables! still need to tune pid and low key kg might be a little off

got feedforward mostly tuned, added a elevator stop method that we call in disabled periodic so now it doesn't jump when it reenables! still need to tune pid and low key kg might be a little off

it works but needs to be retuned. we changed the limit switch stop thing to just set the setpoint to current position but a little away from limit switch (instead of setting voltage to 0), and made it so we only zero at the bottom when we aren't moving up