in theory this is turret code that should work -- needs to be sanity checked before testing
turret sanity check + angleToTicks() function to use in .calculate

fixed unit conversions and tuned pid minimally

worked on understanding the math and tried to get the turret rotation always tracking

addded a supplier for the currebt robot pose in robot container we passed it into the turret subsystem and then used it to make the target angle thing... used that in the periodic to always be at that angle. this is not completely working because of we guessed at the name of the name of the get radians

made it so the code can build

tested changes from last time and not working; added drivetrain code back in

inverted motor, fixed radians vs degrees

uninverted the motor and goes to first angle correctly but no find angle correct once moving, next write test cases

made test case for turret, does not show output

got test cases working, cursor added something to end of build.gradle that made it work

switched over to rotation 2d instead of doubles for the turret angles, had drivetrain problems and tried resetting offsets but it still drives and can't turn in a circle which is sad but decided to just wait for new drivetrain code. turret tracking works!